requirements.txt
requests
pycryptodome
ddddocr
pyexecjs
pymysql
python-dotenv
toml
tabulate
